move ParserOptions to its own module
